backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
cryptography
python-multipart
email-validator
requests
